implement church invitation link
