+++ app/church/[churchId]/invitation.tsx
@@ -1,54 +1,16 @@
+import ChurchHeader from '@/src/components/church/ChurchHeader';
 import LoadingSpinner from '@/src/components/LoadingSpinner';
 import { useChurch } from '@/src/hooks/useChurch';
 import { useAcceptChurchInvite } from '@/src/hooks/useChurchInvitation';
 import { useProfile } from '@/src/hooks/useProfile';
 import ChurchInvitationScreen from '@/src/screens/ChurchInvitationScreen';
 import { useAuth } from '@/src/state/AuthContext';
-import { Ionicons } from '@expo/vector-icons';
 import { Href, Stack, useLocalSearchParams, useRouter } from 'expo-router';
-import { Text, TouchableOpacity, View, useColorScheme } from 'react-native';
-import { useSafeAreaInsets } from 'react-native-safe-area-context';
-
-type ChurchHeaderProps = {
-  title: string;
-  topInset: number;
-  onBackPress: () => void;
-};
-
-function ChurchHeader({ title, topInset, onBackPress }: ChurchHeaderProps) {
-  const colorScheme = useColorScheme();
-  const isDark = colorScheme === 'dark';
-  const backIconColor = isDark ? '#f9fafb' : '#111827';
-
-  return (
-    <View
-      className="border-b border-gray-200 bg-white px-4 pb-3 dark:border-neutral-900 dark:bg-black"
-      style={{ paddingTop: topInset + 8 }}>
-      <View className="relative flex-row items-center justify-center">
-        <TouchableOpacity
-          accessibilityLabel="Go back"
-          accessibilityRole="button"
-          className="absolute left-0 z-10 h-10 w-10 items-center justify-center rounded-full bg-gray-100 dark:bg-neutral-900"
-          onPress={onBackPress}>
-          <Ionicons name="chevron-back" size={22} color={backIconColor} />
-        </TouchableOpacity>
-
-        <View className="min-h-10 max-w-[70%] justify-center">
-          <Text
-            className="text-center text-lg font-semibold text-gray-900 dark:text-white"
-            numberOfLines={1}>
-            {title}
-          </Text>
-        </View>
-      </View>
-    </View>
-  );
-}
+import { Text, TouchableOpacity, View } from 'react-native';
 
 export default function ChurchInvitationRoute() {
   const { churchId, invitedBy } = useLocalSearchParams<{ churchId: string; invitedBy?: string }>();
   const router = useRouter();
-  const insets = useSafeAreaInsets();
   const { session, isGuest } = useAuth();
 
   const churchQuery = useChurch(churchId);
@@ -75,7 +37,7 @@ export default function ChurchInvitationRoute() {
       <>
         <Stack.Screen options={{ headerShown: false }} />
         <View className="flex-1 bg-white dark:bg-black">
-          <ChurchHeader title={headerTitle} topInset={insets.top} onBackPress={handleBackPress} />
+          <ChurchHeader title={headerTitle} onBackPress={handleBackPress} />
           <LoadingSpinner style={{ marginTop: 30 }} />
         </View>
       </>
@@ -87,7 +49,7 @@ export default function ChurchInvitationRoute() {
       <>
         <Stack.Screen options={{ headerShown: false }} />
         <View className="flex-1 bg-white dark:bg-black">
-          <ChurchHeader title={headerTitle} topInset={insets.top} onBackPress={handleBackPress} />
+          <ChurchHeader title={headerTitle} onBackPress={handleBackPress} />
           <View className="flex-1 items-center justify-center px-6">
             <Text className="text-lg font-semibold text-gray-900 dark:text-white">
               Unable to load invitation
@@ -116,7 +78,7 @@ export default function ChurchInvitationRoute() {
     <>
       <Stack.Screen options={{ headerShown: false }} />
       <View className="flex-1 bg-white dark:bg-black">
-        <ChurchHeader title={headerTitle} topInset={insets.top} onBackPress={handleBackPress} />
+        <ChurchHeader title={headerTitle} onBackPress={handleBackPress} />
         <ChurchInvitationScreen
           church={churchQuery.data}
           inviter={inviterQuery.data ?? null}
